Added game updation and token purchase

+++ Auction.sol
@@ -2,7 +2,9 @@
 pragma solidity ^0.8.9;
 import "https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/token/ERC721/IERC721.sol";
 
-contract Auction{
+contract Auction {
+    // not setting winner and transferring eth
+
     // how to automatically call the functions in contracts (to end the auction)
 
     // write events for each one
@@ -11,15 +13,15 @@ contract Auction{
     address payable public owner;
 
     // item struct
-    struct Item{
-        uint id;
+    struct Item {
+        uint256 id;
         // IERC721 nft; // will require an address IERC721(address) to fetch the nft
         string nftName;
-        uint price;
+        uint256 price;
         address payable seller;
         address highestBidder;
-        uint startTime;
-        uint endTime;
+        uint256 startTime;
+        uint256 endTime;
         State state;
         address winner;
     }
@@ -31,42 +33,40 @@ contract Auction{
     }
 
     // to store the bidder info
-    struct Bidder{
+    struct Bidder {
         address payable bidderAddress;
-        uint bidAmount;
+        uint256 bidAmount;
     }
 
     // to store all the items
     Item[] items;
 
     // to maintain the index
-    uint index;
+    uint256 index;
 
     // to get individual Item info from ID
-    mapping(uint => Item) public getItemDetails;
+    mapping(uint256 => Item) public getItemDetails;
 
     // id => all the bidders
-    mapping(uint => Bidder[]) getBidders; 
+    mapping(uint256 => Bidder[]) getBidders;
 
     // get highest bidder info
     mapping(address => Bidder) getHighestBidder;
 
     // constructor to set owner
-    constructor(){
+    constructor() {
         owner = payable(msg.sender);
     }
 
     // create auction
     function create(
-        uint _id,
+        uint256 _id,
         // address _nft,
         string memory _nftName,
-        uint _price,
-        uint _startTime,
-        uint _endTime
-    )
-    public{
-
+        uint256 _price,
+        uint256 _startTime,
+        uint256 _endTime
+    ) public {
         // manager can not create - make this reverse (bija koi no kari sakva joiye)
         require(msg.sender == owner, "Only Owner can create");
 
@@ -99,15 +99,18 @@ contract Auction{
         index++;
     }
 
-    function start(uint _id) public {
+    function start(uint256 _id) public {
         // only owner can start
         require(msg.sender == owner, "Only owner can start auction");
 
         // needs to be in created state
-        require(getItemDetails[_id].state == State.Created , "Not in Created State");
+        require(
+            getItemDetails[_id].state == State.Created,
+            "Not in Created State"
+        );
 
         // need to be valid time
-        // require(block.timestamp == getItemDetails[_id].startTime || 
+        // require(block.timestamp == getItemDetails[_id].startTime ||
         // (
         //     block.timestamp > getItemDetails[_id].startTime && block.timestamp < getItemDetails[_id].endTime
         // ),
@@ -118,12 +121,15 @@ contract Auction{
         getItemDetails[_id].state = State.Ongoing;
     }
 
-    function bid(uint _id, uint _amount) public{
+    function bid(uint256 _id, uint256 _amount) public {
         // owner can  not bid
         require(msg.sender != owner, "owner can not bid");
 
         // needs to be in ongoing state
-        require(getItemDetails[_id].state == State.Ongoing , "Not in Ongoing State");
+        require(
+            getItemDetails[_id].state == State.Ongoing,
+            "Not in Ongoing State"
+        );
 
         // getting item with the given id
         Item storage _selectedItem = getItemDetails[_id];
@@ -135,10 +141,16 @@ contract Auction{
         Bidder storage _bidder = getHighestBidder[currHightest];
 
         // bid should be greater than the price
-        require(_amount > _selectedItem.price, "Your bid needs to be higher than the price");
+        require(
+            _amount > _selectedItem.price,
+            "Your bid needs to be higher than the price"
+        );
 
         // comparing amounts
-        require(_amount > _bidder.bidAmount, "Your bid needs to be higher than the current bid"); 
+        require(
+            _amount > _bidder.bidAmount,
+            "Your bid needs to be higher than the current bid"
+        );
 
         // setting the new highest bidder
         _selectedItem.highestBidder = msg.sender;
@@ -147,8 +159,7 @@ contract Auction{
         getBidders[_id].push(Bidder(payable(msg.sender), _amount));
     }
 
-    function end(uint _id) public payable{
-
+    function end(uint256 _id) public payable {
         // only owner can end
         require(msg.sender == owner, "Only owner can end");
 
@@ -162,27 +173,29 @@ contract Auction{
         Bidder storage _bidder = getHighestBidder[currHightest];
 
         // checking the state
-        require(_selectedItem.state == State.Ongoing, "The state needs to be Ongoing");
+        require(
+            _selectedItem.state == State.Ongoing,
+            "The state needs to be Ongoing"
+        );
 
         // time should be right
         // require(_selectedItem.endTime <= block.timestamp, "You can not end the auction yet");
 
         // fund transfer
         _selectedItem.seller.transfer(_bidder.bidAmount);
-        
+
         _selectedItem.winner = _bidder.bidderAddress;
 
         // set the state to ended
         _selectedItem.state = State.Ended;
-
     }
 }
 
 // // transfer the funds if the current highest bidder has the amount available
 //         if(currHightest.balance >= _bidder.bidAmount){
-            
+
 //             _selectedItem.seller.transfer(_bidder.bidAmount);
-        
+
 //             // set the state to ended
 //             _selectedItem.state = State.Ended;
 
@@ -195,14 +208,14 @@ contract Auction{
 //             Bidder[] memory _biddersArray = getBidders[_id];
 
 //             for(uint i = 0; i < _biddersArray.length; i++){
-                
+
 //                 Bidder memory tempBidder = getHighestBidder[_biddersArray[i].bidderAddress];
 
 //                 if(tempBidder.bidderAddress.balance >= tempBidder.bidAmount){
 
 //                     _selectedItem.seller.transfer(tempBidder.bidAmount);
 //                     _selectedItem.winner = tempBidder.bidderAddress;
-    
+
 //                     // set the state to ended
 //                     _selectedItem.state = State.Ended;
